Fix unloading the extension
Since GNOME 3.14, the clock label in the panel has been updated to use a
property binding to the wall clock instance, instead of using a function
on the DateMenuButton object. This means that unloading the extension
started failing when the disable() method gets called and tries to
replace a non-existing function.

Now we use a proper signal connection that we can keep around and toggle
at will.

This means we're going to drop backward compatibility for all Shell
versions before 3.14, which was released September 2014.